no warnings on MSVC

+++ src/ast/ast_allocate_stack.cpp
@@ -164,7 +164,7 @@ namespace das {
     // New
         virtual ExpressionPtr visit ( ExprNew * expr ) override {
             if ( expr->type->dim.size() ) {
-                auto sz = expr->type->getCountOf() * sizeof(char *);
+                auto sz = uint32_t(expr->type->getCountOf()*sizeof(char *));
                 expr->stackTop = allocateStack(sz);
                 if ( log ) {
                     logs << "\t" << expr->stackTop << "\t" << sz
